+++ components/result_table_filter.go
@@ -22,7 +22,7 @@ func NewResultsFilter() *ResultsTableFilter {
 		Input: tview.NewInputField(),
 		Label: tview.NewTextView(),
 	}
-	recordsFilter.SetBorder(true)
+	// recordsFilter.SetBorder(true)  // Remove border to save space
 	recordsFilter.SetDirection(tview.FlexRowCSS)
 	recordsFilter.SetTitleAlign(tview.AlignCenter)
 	recordsFilter.SetBorderPadding(0, 0, 1, 1)